+++ packages/nestjs/src/index.ts
@@ -1,0 +1,7 @@
+export { ConfigBoundModule } from './config-bound.module';
+export { ConfigBoundService } from './config-bound.service';
+export {
+  ConfigBoundModuleOptions,
+  ConfigBoundModuleAsyncOptions,
+  ConfigBoundOptionsFactory
+} from './interfaces/config-bound-module-options.interface';